java

+++ BasicCalculator/src/main/java/org/example/Main.java
@@ -9,10 +9,10 @@ public class Main {
         System.out.println("Enter the first number:");
         // use scanner to allow user to write input
         Scanner scanner = new Scanner(System.in);
-        float firstNumber = scanner.nextFloat();
+        double firstNumber = scanner.nextDouble();
         // asks for second number
         System.out.println("Enter the second number:");
-        float secondNumber = scanner.nextFloat();
+        double secondNumber = scanner.nextDouble();
         // prints the possible calculations and asks user to enter an option
         System.out.println("Possible Calculations:");
         System.out.println("(A)dd");
@@ -23,19 +23,19 @@ public class Main {
         String answer = scanner.next();
         // if statement for addition
         if (answer.equalsIgnoreCase("A")) {
-            float sum = firstNumber + secondNumber;
+            double sum = firstNumber + secondNumber;
             System.out.printf("%.2f + %.2f = %.2f", firstNumber, secondNumber, sum);
         }   // else if for subtraction
         else if (answer.equalsIgnoreCase("S")) {
-            float subtract = firstNumber - secondNumber;
+            double subtract = firstNumber - secondNumber;
             System.out.printf("%.2f - %.2f = %.2f", firstNumber, secondNumber, subtract);
         }   // else if for multiplication
         else if (answer.equalsIgnoreCase("M")) {
-            float multiply = firstNumber * secondNumber;
+            double multiply = firstNumber * secondNumber;
             System.out.printf("%.2f * %.2f = %.2f", firstNumber, secondNumber, multiply);
         }   // else if for division
         else if (answer.equalsIgnoreCase("D")) {
-            float divide = firstNumber / secondNumber;
+            double divide = firstNumber / secondNumber;
             System.out.printf("%.2f / %.2f = %.2f", firstNumber, secondNumber, divide);
         }   // else for an input that's not a,s,m,d
         else {
